separate subscription and auth flow
